Move config bootstrap into common lib
