+++ trac-libs/trac-lib-common/src/main/java/com/accenture/trac/common/config/ConfigBootstrap.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2021 Accenture Global Solutions Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.accenture.trac.common.config;
+
+import com.accenture.trac.common.util.VersionInfo;
+
+
+public class ConfigBootstrap {
+
+    public static ConfigManager useCommandLine(Class<?> serviceClass, String[] args) {
+
+        var componentName = VersionInfo.getComponentName(serviceClass);
+        var componentVersion = VersionInfo.getComponentVersion(serviceClass);
+
+        var startupBanner = String.format(">>> %s %s", componentName, componentVersion);
+        System.out.println(startupBanner);
+
+        var standardArgs = StandardArgsProcessor.processArgs(componentName, args);
+
+        System.out.println(">>> Working directory: " + standardArgs.getWorkingDir());
+        System.out.println(">>> Config file: " + standardArgs.getConfigFile());
+        System.out.println();
+
+        var configManager = new ConfigManager(standardArgs);
+        configManager.initConfigPlugins();
+        configManager.initLogging();
+
+        return configManager;
+    }
+}
